chore: bump alloy evm and revm deps (#515)

+++ crates/dora-runtime/src/vm.rs
@@ -239,7 +239,7 @@ impl<DB: Database> VM<DB> {
         blobs: &[B256],
         max_blob_fee: u128,
         block_blob_gas_price: u128,
-        max_blobs: u8,
+        max_blobs: u64,
     ) -> Result<(), VMError> {
         // Ensure that the user was willing to at least pay the current blob gasprice
         if block_blob_gas_price > max_blob_fee {
